reorganize big file into require()able modules
