+++ src/js/config/hrefWhitelist.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = function(app) {
+  // allow data: hrefs
+  app.config(['$compileProvider', function($compileProvider) {
+    $compileProvider.aHrefSanitizationWhitelist(/^\s*(https?|mailto|data):/);
+  }]);
+};
