+++ src/js/config/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = function(app) {
+  require('./hrefWhitelist')(app);
+  require('./translate')(app);
+};
