+++ src/js/directives/compareTo.js
@@ -1,0 +1,23 @@
+'use strict';
+
+module.exports = function(app) {
+  // http://odetocode.com/blogs/scott/archive/2014/10/13/confirm-password-validation-in-angularjs.aspx
+  app.directive('compareTo', function() {
+    return {
+      require: 'ngModel',
+      scope: {
+        otherModelValue: '=compareTo'
+      },
+      link: function(scope, element, attributes, ngModel) {
+
+        ngModel.$validators.compareTo = function(modelValue) {
+          return modelValue == scope.otherModelValue;
+        };
+
+        scope.$watch('otherModelValue', function() {
+          ngModel.$validate();
+        });
+      }
+    };
+  });
+};
