+++ src/js/directives/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = function(app) {
+  require('./compareTo')(app);
+};
