+++ src/js/filters/base64encode.js
@@ -1,0 +1,9 @@
+'use strict';
+
+module.exports = function(app) {
+  app.filter('base64encode', ['$window', function($window) {
+    return function(input) {
+      return $window.btoa(input);
+    };
+  }]);
+};
