+++ src/js/filters/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = function(app) {
+  require('./base64encode')(app);
+  require('./range')(app);
+};
