+++ src/js/services/index.js
@@ -1,5 +1,6 @@
 'use strict';
 
-module.exports = function(module) {
-  require('./jsonrpc')(module);
+module.exports = function(app) {
+  require('./downloadFile')(app);
+  require('./jsonrpc')(app);
 };
